nrepl functional redesign
Changed up the design and layout of how messages are constructed and
used. Deleted the message modules.

Changed to use a fork of the Bento library that allows for partial
parsing. This was important to get a functional bencode response
reader.

Changed the connection to interact with the tcp socket not in :active
mode and return a Stream of bencode objects.

Deleted most of the tests and will work on making new ones.

Added a usage block to the README to make it very clear how one could
start using the library.